Move shared load graph code
