+++ packages/graph-explorer/src/utils/index.ts
@@ -16,3 +16,4 @@ export * from "./set";
 export * from "./env";
 export * from "./constants";
 export * from "./NetworkError";
+export * from "./formatEntityCounts";
